+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Simple Type Caster Change Log
 
+## 0.4.3 December 15, 2024
+
+- Add `TypeCaster::toInt()`.
+- Add `TypeCaster::toDateTimeOrNullByTimestamp`.
+- Refactor `TypeCaster::toIntOrNull()`.
+
 ## 0.4.2.1 December 15, 2024
 
 - Fix `TypeCaster::toString()` when value is not scalar and not array.